Test Kernel and mocks refactored to use additional passed DI definitions.
