+++ tests/ITBShopwareSdkBundleKernel.php
@@ -13,6 +13,7 @@ use ITB\ShopwareSdkBundle\Tests\Compiler\PsrSimpleCacheServicesCompilerPass;
 use ITB\ShopwareSdkBundle\Tests\Compiler\PublishServicesForTestsCompilerPass;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\HttpKernel\Bundle\BundleInterface;
 use Symfony\Component\HttpKernel\Kernel;
 
@@ -27,12 +28,19 @@ class ITBShopwareSdkBundleKernel extends Kernel
     private ?array $shopwareSdkConfig;
 
     /**
-     * @param ITBShopwareSdkConfiguration|null $shopwareSdkConfig
+     * @var array<string, Definition>
      */
-    public function __construct(string $environment, bool $debug, ?array $shopwareSdkConfig = null)
+    private array $additionalDefinitions;
+
+    /**
+     * @param ITBShopwareSdkConfiguration|null $shopwareSdkConfig
+     * @param array<string, Definition> $additionalDefinitions
+     */
+    public function __construct(string $environment, bool $debug, ?array $shopwareSdkConfig = null, array $additionalDefinitions = [])
     {
         parent::__construct($environment, $debug);
         $this->shopwareSdkConfig = $shopwareSdkConfig;
+        $this->additionalDefinitions = $additionalDefinitions;
     }
 
     public function getCacheDir(): string
@@ -55,11 +63,15 @@ class ITBShopwareSdkBundleKernel extends Kernel
                 $container->loadFromExtension(ITBShopwareSdkExtension::ALIAS, $this->shopwareSdkConfig);
             }
 
+            foreach ($this->additionalDefinitions as $id => $definition) {
+                $container->setDefinition($id, $definition);
+            }
+
             $container->addCompilerPass(new PsrHttpServicesCompilerPass());
             $container->addCompilerPass(new PsrClockServicesCompilerPass());
             $container->addCompilerPass(new PsrSimpleCacheServicesCompilerPass());
             // All services are made public to use them via container.
-            $container->addCompilerPass(new PublishServicesForTestsCompilerPass());
+            $container->addCompilerPass(new PublishServicesForTestsCompilerPass(), priority: -1000);
         });
     }
 }
