+++ tests/Mock/AdditionalDefinitionCollectionPopulator.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ITB\ShopwareSdkBundle\Tests\Mock;
+
+use Vin\ShopwareSdk\Definition\DefinitionCollection;
+use Vin\ShopwareSdk\Definition\DefinitionCollectionPopulator;
+
+final class AdditionalDefinitionCollectionPopulator implements DefinitionCollectionPopulator
+{
+    public static function priority(): int
+    {
+        return 1;
+    }
+
+    public function populateDefinitionCollection(DefinitionCollection $definitionCollection, string $shopwareVersion): void
+    {
+        $definitionCollection->set(new AdditionalDefinition());
+    }
+}
